changed limits to STEREO

+++ scripts/hv_stereo_pull2_new.py
@@ -32,6 +32,10 @@ class URLLister(SGMLParser):
                         self.urls.extend(href)
 
 
+def checkFileSize(f,minsize):
+	stat = os.stat(f)
+	return stat.st_size > minsize
+
 # Create a time-stamp to be used by all log files
 def createTimeStamp():
 	TSyyyy = time.strftime('%Y',time.localtime())
@@ -315,9 +319,12 @@ def GetAIAWave(nickname,yyyy,mm,dd,wave,remote_root,local_root,ingest_root,monit
 					for name in newlist:
 						newFile = name[:-1]
 						if newFile.endswith('.jp2'):
-							shutil.copy2(local_keep + newFile,moveTo + newFile)
-							doJPIPencoding.doJPIPencoding(moveTo + newFile,'SOHO')
-							change2hv(moveTo + newFile)
+							if checkFileSize(local_keep + newFile,minJP2SizeInBytes):
+								shutil.copy2(local_keep + newFile,moveTo + newFile)
+								doJPIPencoding.doJPIPencoding(moveTo + newFile,'SOHO')
+								change2hv(moveTo + newFile)
+							else:
+								jprint(local_keep + newFile+' is smaller than minimum allowed.')
 					#if os.path.exists(os.path.expanduser(local_keep + newFile)):
 					#	os.remove(local_keep + newFile)
 			else:
@@ -442,6 +449,7 @@ else:
 	daysBackMin = int(options[11][:-1])
 	daysBackMax = int(options[12][:-1])
 
+
 	# Re-direct stdout to a logfile?
 	if redirectTF == 'True':
 		redirect = True
@@ -461,8 +469,8 @@ else:
 		count = 0
 		while 1:
 			count = count + 1
-			print '********************'
-			print count
+			#print '********************'
+			#print count
 			gotNewData = False
 			for daysBack in range(daysBackMin,daysBackMax):
